requirements.txt
PyJWT~=2.6.0
Flask~=2.2.2